+++ mycodo/inputs/atlas_do.py
@@ -1,11 +1,18 @@
 # coding=utf-8
 import copy
+import time
 
 from flask_babel import lazy_gettext
 
+from mycodo.databases.models import Conversion
 from mycodo.inputs.base_input import AbstractInput
+from mycodo.inputs.sensorutils import convert_from_x_to_y_unit
+from mycodo.utils.atlas_calibration import AtlasScientificCommand
 from mycodo.utils.atlas_calibration import setup_atlas_device
 from mycodo.utils.constraints_pass import constraints_pass_positive_value
+from mycodo.utils.database import db_retrieve_table_daemon
+from mycodo.utils.system_pi import get_measurement
+from mycodo.utils.system_pi import return_measurement_info
 from mycodo.utils.system_pi import str_is_float
 
 # Measurements
@@ -51,12 +58,24 @@ INPUT_INFORMATION = {
 
     'custom_options': [
         {
+            'id': 'temperature_comp_meas',
+            'type': 'select_measurement',
+            'default_value': '',
+            'options_select': [
+                'Input',
+                'Function',
+                'Math'
+            ],
+            'name': "{}: {}".format(lazy_gettext('Temperature Compensation'), lazy_gettext('Measurement')),
+            'phrase': lazy_gettext('Select a measurement for temperature compensation')
+        },
+        {
             'id': 'max_age',
             'type': 'integer',
             'default_value': 120,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
-            'name': "{}: {}".format(lazy_gettext('Calibrate'), lazy_gettext('Max Age')),
+            'name': "{}: {}".format(lazy_gettext('Temperature Compensation'), lazy_gettext('Max Age')),
             'phrase': lazy_gettext('The maximum age (seconds) of the measurement to use')
         }
     ],
@@ -108,8 +127,10 @@ class InputModule(AbstractInput):
 
         self.atlas_device = None
         self.interface = None
-        self.calibrate_sensor_measure = None
+        self.atlas_command = None
 
+        self.temperature_comp_meas_device_id = None
+        self.temperature_comp_meas_measurement_id = None
         self.max_age = None
 
         if not testing:
@@ -119,7 +140,6 @@ class InputModule(AbstractInput):
 
     def initialize_input(self):
         self.interface = self.input_dev.interface
-        self.calibrate_sensor_measure = self.input_dev.calibrate_sensor_measure
 
         try:
             self.atlas_device = setup_atlas_device(self.input_dev)
@@ -129,6 +149,10 @@ class InputModule(AbstractInput):
             self.atlas_device.atlas_write("O,%,0")
             ret_val = self.atlas_device.atlas_write("O,?")
             self.logger.info("Parameters enabled: {}".format(ret_val))
+
+            if self.temperature_comp_meas_measurement_id:
+                self.atlas_command = AtlasScientificCommand(
+                    self.input_dev, sensor=self.atlas_device)
         except Exception:
             self.logger.exception("Exception while initializing sensor")
 
@@ -143,6 +167,38 @@ class InputModule(AbstractInput):
 
         do = None
         self.return_dict = copy.deepcopy(measurements_dict)
+
+        # Compensate measurement based on a temperature measurement
+        if self.temperature_comp_meas_measurement_id and self.atlas_command:
+            self.logger.debug("DO sensor set to calibrate temperature")
+
+            device_measurement = get_measurement(
+                self.temperature_comp_meas_measurement_id)
+            conversion = db_retrieve_table_daemon(
+                Conversion, unique_id=device_measurement.conversion_id)
+            channel, unit, measurement = return_measurement_info(
+                device_measurement, conversion)
+
+            last_measurement = self.get_last_measurement(
+                self.temperature_comp_meas_device_id,
+                self.temperature_comp_meas_measurement_id,
+                max_age=self.max_age)
+
+            out_value = convert_from_x_to_y_unit(unit, "C", last_measurement[1])
+
+            if last_measurement:
+                self.logger.debug(
+                    "Latest temperature used to calibrate: {temp}".format(
+                        temp=out_value))
+                ret_value, ret_msg = self.atlas_command.calibrate(
+                    'temperature', set_amount=out_value)
+                time.sleep(0.5)
+                self.logger.debug("Calibration returned: {val}, {msg}".format(
+                    val=ret_value, msg=ret_msg))
+            else:
+                self.logger.error(
+                    "Calibration measurement not found within the past {} seconds".format(
+                        self.max_age))
 
         # Read sensor via FTDI or UART
         if self.interface in ['FTDI', 'UART']: