+++ mycodo/inputs/dht11.py
@@ -1,13 +1,10 @@
 # coding=utf-8
+import copy
 import time
 
-import copy
-
-from mycodo.databases.models import Output
 from mycodo.inputs.base_input import AbstractInput
 from mycodo.inputs.sensorutils import calculate_dewpoint
 from mycodo.inputs.sensorutils import calculate_vapor_pressure_deficit
-from mycodo.utils.database import db_retrieve_table_daemon
 
 # Measurements
 measurements_dict = {